requirements.txt
Flask==2.1.2
Jinja2==3.1.2
gunicorn==20.1.0
line-bot-sdk==2.2.1
openai==0.27.2
python-dotenv==1.0.0
